ML/requirements.txt
# ============================================
# Best Time Post API - Requirements
# Python 3.11+
# ============================================

# --- API Framework ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# --- Configuration & Validation ---
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# --- Rate Limiting & Caching ---
slowapi==0.1.9
cachetools==5.3.2

# --- HTTP Clients ---
requests==2.31.0
httpx==0.25.2

# --- Apify Data Collection ---
apify-client==1.6.0

# --- Database ---
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# --- Data Processing & Analysis ---
pandas==2.1.4
numpy==1.26.2

# --- Machine Learning ---
scikit-learn==1.3.2
joblib==1.3.2
lightgbm==4.1.0

# --- Deep Learning (PyTorch) ---
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# --- Neural Prophet ---
neuralprophet==0.7.0
prophet==1.1.5

# --- ONNX Export & Inference ---
onnx==1.15.0
onnxruntime==1.16.3
onnxmltools==1.11.2
skl2onnx==1.16.0

# --- Time Series & Forecasting ---
statsmodels==0.14.1

# --- Logging ---
loguru==0.7.2

# --- Time & Date ---
pytz==2023.3.post1
python-dateutil==2.8.2

# --- Visualization (Optional, for notebooks) ---
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# --- Jupyter (Optional, for notebooks) ---
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1
ipywidgets==8.1.1

# --- Development & Testing (Optional) ---
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# --- Utilities ---
tqdm==4.66.1
click==8.1.7
